Release 2.7.1-beta.9 to pre-release users
This change releases the new beta (2.7.1.9) to the users who have selected to try the pre-release versions. They will currently be using v2.5.0, and this will upgrade them to v2.7.1.9.

I need to release the final version as 2.7.2.0 so that they will be automatically upgraded as the next version (as 2.7.1.10 would not upgrade correctly).