Added multiple profile apply reruns to workaround some slow monitors
Some monitors like the Samsung Odessey G9 don't accept the display profile setting until the second time we apply them. This change builds the settings into the Profiles.json file to allow users to set the number of times the profile is applied, and the delay between apply attempts in milliseconds.