[WIP] Added Apply Profile multiple times to Profile Settings Form
Added the basic UI to the Profile Settings form in preparation for wiring it up to the Profile Item objects. This is the 2nd step for adding this functionality to DisplayMagician.

Next and final step is to wire up the form functions to prepopulate the UI with the value from the Profile Item, and to ensure that the profile item is updated and saved when the UI changes.