Added ability to apply profile multiple times
You need to go into the Profile Settings to set the number of times that the profile is applied, and how long the delay should be in seconds before it is reapplied each time. Fixes #316 (well it provides a workaround for #316).

Note - The target display profile you have selected in the game shortcut will be applied as per the number of times that it is set in the target display profile, but the profile that is reverted back to will only ever be applied once! This is because the display profile that you start with is a temporary display profile that is only stored whilst the game is being run. It is discarded when the game shortcut has finished processeing and everything is reverted back.

A second change I made was to extend the default timeout for waiting for a game or exe to start to be 60 seconds. It hwas fairly common that a game would take too long to start, and it would timeout and DM would revert the display, only for the game to then finish starting (but the display profile would have reverted back). Extending the timeout fixes that issue.

Lastly, a few Task.Delays were incorrectly being used, as they were not pausing properly. I added the .Wait(cancelToken) to them so that they would properly delay, and yet would be cancellable so that the user could cancel if they wanted.