[WIP] Partially moved Game scanning code to ShortcutLibraryForm
Have moved game library scanning code and app library scanning code to the ShortcutLibraryForm edit button to ensure that the game library is as fast loading as possible. Some users have very very large lists of games, and that made it very slow for DM to load.

This partial changes makes DM faster 90% of the time, and its only when editing a shortcut or creating a new one does it go too slow.

This version only works for Program.AppAppList for the place to store the list of known apps. The Program.AppGameList part is broken :(. That needs fixing.