Later game loading until adding or editing shortcut
This speeds up DisplayMagician greatly for normal use. DM now only inspects the apps and games installed when you edit or create a new Game Shortcut. This can take minutes for people with extensive game collectiions. This change means that everything loads faster for 90% of the time, and the last 10% is fine. Fixes #373.

Also changed ShortcutLibrary and ShortcutForm to center on the Parent form, so that it's easier for users to use.