Preparation for allowing users to disable WakeGPUs setting
This wakeGPUs setting will eventually keep the GPUS all awake to prevent the GPUs from sleeping. If enabled this setting will ensure that all GPPUS (and therefore all displays connected to all GPUs) are reachable and therefore can be used. This will increase battery use in gaming laptops, so that is why I have added this checkbox to turn stuff off.

I have not actually implemented the setting so it works though, as its only the latest video libraries that I've made that support this feature, but those haven't finished being tested in VideoInfo as yet. So the rest of this work will need to wait until that part is finished.