Fixed mainForm visibility logic when different settings are active
The mainForm was not opening properly when the minimise was not set. This error was caused by logic changes when I main the MinForm always run (even when hidden) in order to mandate a proper hierarchy for the windows. This now has been fixed.

Additionally the main loading splashscreen has been re-enabled so that users can easily see that DM is loading even when it is minimised. This can be changed in the program settings for those users who want a completely hidden offering.